requirements.txt
Django==2.2
djangorestframework==3.12.4
Pillow==8.2.0
docutils==0.17.1
drf-yasg==1.20.0